minor fixups

+++ src/scenes/boids/boidRules.hpp
@@ -4,9 +4,9 @@
 class BoidRules {
 public:
 	BoidRules(): separationWeight(2.0), cohesionWeight(0.25), alignmentWeight(0.5), desiredSeparation(100), cohesionDist(150), alignmentDist(200) {;}
-	BoidRules(int separation): desiredSeparation(separation), separationWeight(2.0), cohesionWeight(0.25), alignmentWeight(0.5), cohesionDist(150), alignmentDist(200) {;} // TODO testing
+	BoidRules(float separation): desiredSeparation(separation), separationWeight(2.0), cohesionWeight(0.25), alignmentWeight(0.5), cohesionDist(150), alignmentDist(200) {;} // TODO testing
 
-	friend std::ostream& operator<<(std::ostream &strm, const BoidRules &br) { // TODO flesh out
+	friend std::ostream& operator << (std::ostream &strm, const BoidRules& br) { // TODO flesh out
 		return strm << "BoidRules: sw: " << br.separationWeight << ", ds: " << br.desiredSeparation << "\n";
 	}
